J# 33089 J#33090 J# 33091 J# 33092 J# 33093
Fixed the codesystem displayed name for AttributeEstimateType in the Evidence resource. attributeEstimate is now in modelCharacteristic as a linking. And attributeEstimate.attributeEstimate now links to attributeEstimate. Fixed and updated a few valueset definitions for Evidence and EvidenceVariable.

Evidence Resource was updated to add statistic.modelCharacteristic.attributeEstimate
Citation Resource was updated to:
o change the JournalIssueMedium value set to CitedMedium value set and revise the 6 terms in this value set, and
o to add a coverage term to the CitedArtifactClassificationType value set, and
o to revise the 7 terms in the ContributorSummaryType value set, and
o to revise the 7 terms in the ContributorRole value set