J# 36779 CodeSystem Characteristic Combination changes
When modeling the use of EvidenceVariable Resource to define evidence variables used in meta-analyses, we discovered two additional types of combinations of characteristics that are not supported by the required code set (all-of, any-of, at-least, at-most, net-effect).

A 'statistical' code is needed when the Evidence Variable is defined as a statistical combination of other characteristics.  The code 'statistical' can have display 'Statistical' and definition 'Combine characteristics statistically. Use method to specify the statistical method.'

The definition for the code 'net-effect' will need to change from 'Combine characteristics statistically.' to 'Combine characteristics by addition of benefits and subtraction of harms.'

A 'dataset' code is needed when the Evidence Variable is defined as a collection of other characteristics where each characteristic is an observed value used in the statistical analysis. In modeling a meta-analysis this may be achieved by referencing an Evidence Resource as the characteristic.definitionReference.  The code 'dataset' can have display 'Dataset' and definition 'Combine characteristics as a collection used as the dataset.'