Additional changes for MolSeq for referenceseq and only allowing a single representation
Change the binding for the referenceSeqId, and remove the local valueset (it isn't best practice to allow a valueset to bring in overlapping codes from different code systems).
Implement a constraint to only allow a single sequence representation (literal, formatted, or relative). Adjusted guidance to indicate this more clearly.
Updated chromsome binding to a more commonly used LOINC answer list.