J# 36709 and J# 32314
For ClinicalTrials.gov use, added codes to ResearchStudyStatus codesystem:  for not-yet-recruiting, recruiting, enrolling-by-invitation, active-but-not-recruiting, and terminated.

Add the new element associatedParty.period with cardinality of 0..*
Datatype would be Period;
Description:
Short Def. - when active in the role;
Long Def. - Identifies the start date and the end date of the associated party in the role.
Comment: the cardinality is 0..* due to the fact that an associated party may be intermittently active in a given role over multiple time periods.