FHIR-19620 Clarify patch with history
Applied the resolution items to update the language pertaining to how history entries are created and how the use of patch interaction is represented as an update interaction in the history Bundle. Additionally, added a new invariant (constraint) bld-14 to the Bundle resource to enforce that the entry.request.method PATCH cannot be used when type = 'history'.