J# 37730 and J# 37731 ResearchStudy resource changes
Definition changes for a couple elements in ResearchStudy

ResearchStudy.title
Definition: The human readable name of the research study.
Short description: Human readable name of the study.

ResearchStudy.date
Definition: The date (and optionally time) when the title was published. The date must change when the business version changes and it must change if the status code changes. In addition, it should change when the substantive content of the title changes.