J# 42886 Replaces EvidenceVariable.characteristic with EvidenceVariable.definition
Group Resource is now changed to be the preferred Resource for defining combinations of characterstics and the characteristic BackboneElement in EvidenceVariable is no longer the preferred method.

The EvidenceVariable definition 0..1 element will have a CodeableReference datatype.

When definition.concept is used, there is no restriction on the value set binding. If an Example-bound value set is required in the structure definition, then example terms can include http://snomed.info/sct#182886004 Placebo given (situation) and http://snomed.info/sct#419620001 Death (event).

When definition.reference is used, the Reference type shall be limited to Group.

Remove EvidenceVariable.characteristic (rather than deprecate, on the grounds that there are no signficant implementations that would need deprecation rather than the preferred approach of removing the element altogether)

Modify the EvidenceVariable introductory text to match the new structure.

We currently update one example, and removed the rest The removed examples will be updated and come back in future commits.