J#48855 Improved Citation per CDS R6 Review Checklist
The following changes were needed to Citation, based on CDS R6 Review Checklist:

1. for Citation.version

change the Note to read: This is a business version label used for display in understanding the formal version of the citation record, not a resource version id or technical version number for an archival system.
2. Add Citation.recorder 0..* ContactDetail

short/definition - Who entered the data for the citation record
comments = The Citation.recorder element is structured to support a name (with a string datatype) and contact point information for the recorder, and is not structured to support unique identifiers for the recorder. If detailed contributorship data is needed, then one can create a Citation Resource for the Citation Resource.
3. for Citation.citedArtifact.relatedIdentifier

comment - add 'a' in 'use as unique identifier'
4. for Citation.citedArtifact.version

change the Note to read: This is a business version label used for display in understanding the formal version of the cited artifact, not a resource version id or technical version number for an archival system.
change datatype to string
5. delete Citation.citedArtifact.version.value

6. delete Citatoin.citedArtifact.version.baseCitation

7. delete Citation.citedArtifact.part.baseCitation

8. add Citation.citedArtifact.baseCitation 0..1 Reference (Citation)

short/definition = Citation for the primary version and complete form of the cited artifact.
comment = When referencing a baseCitation, one may inherit any data from the referenced Citation Resource and consider any data in the current Citation Resource as superceding to be version-specific or part-specific. For example, the part-specific Citation may have additional contributorship data that supercedes the contributorship data for the full citation when there is a need to uniquely report the contributorship for an image or supplement associated with the fuller citation record.
9. for Citation.citedArtifact.publicationForm

add to definition:  Where, when and how the artifact was published.
10. for Citation.citedArtifact.publicationForm.publicationDateSeason

Definition - change to "Season in which the cited artifact was published, e.g. Spring, Summer, Fall/Autumn, Winter."
11. for Citation.citedArtifact.contributorship.complete

change short to 'Indicates if all authors are included'
change definition to 'Indication whether all authors are included in the entry content.'
12. for Citation.citedArtifact.contributorship.entry.contributionType

from definition remove "This element identifies"
13. for Scope and Usage

change "to current work" to "to current or prior work"
add comma after "(such as searching MEDLINE)"
change "Standardization of citations" to "Standardized expression of citations"
capitalize the R in Citation resource, and FHIR resources
add comma in "other FHIR resources, including"
add software to the list with "database updates, and the like."
change "Within citedArtifact.publicationForm, the publishedIn element identifies the largest container (journal, book, database, or newspaper), the periodicRelease element identifies the periodic issue, and the remaining elements (covering dates, language, and pagination) are specific to the artifact being cited." to "Within citedArtifact.publicationForm, the publishedIn element identifies the container (journal, book, database, or newspaper), and the remaining elements (covering dates, language, and pagination) are specific to the artifact being cited."
14. fir Boundaries and Relationships

change "there is additional needs" to "there are additional needs"
15. for 5Ws - status and author are direct matches

16. Citation.citedArtifact.abstract.type should change from 0..1 to 0..*

17. Citation.citedArtifact.contributorship.summary.style should change from 0..1 to 0..*

18. Change the following bindings from Extensible to Preferred:

Citation Summary Style
Citation Classification Type
Cited Artifact Classification Type
Artifact Contribution Instance Type
Contributor Summary Source
19. Change the following binding from Extensible to Example:

Contributor Summary Style
20. Apply Meaning if missing: Value unknown to each of the following boolean elements:

Citation.statusDate.actual - Meaning if missing: Value unknown
Citation.citedArtifact.statusDate.actual - Meaning if missing: Value unknown
Citation.citedArtifact.contributorship.complete - Meaning if missing: Value unknown
Citation.citedArtifact.contributorship.entry.correspondingContact - Meaning if missing: Value unknown
21. add isSummary to approvalDate and lastReviewDate