J# 49755 Improve EvidenceVariable per CDS R6 Review Checklist
1. for EvidenceVariable.status - in the isModifier, change bracketed {title} to EvidenceVariable

2. for EvidenceVariable.definition - change the definition to "The meaning of the evidence variable." (Changed this for short and removed the definition)

3. Add EvidenceVariable.recorder
cardinality 0..*
type ContactDetail
short = Who entered the data for the evidence variable
definition = The person or entity that entered the data into the EvidenceVariable Resource instance, if different than the author or creator of the intellectual property contained within the EvidenceVariable Resource instance.
comments = The EvidenceVariable.recorder element is structured to support a name (with a string datatype) and contact point information for the recorder, and is not structured to support unique identifiers for the recorder. If detailed contributorship data is needed, then one can create a Citation Resource for the EvidenceVariable Resource.

4. for EvidenceVariable.description - change definition to "A natural language summary of the evidence variable."

5. for EvidenceVariable.copyright
change short to “intellectual property ownership, may include restrictions on use”
change definition to “A copyright statement relating to the EvidenceVariable and/or its contents. Copyright statements are legal notices of intellectual property ownership and may include restrictions on the use and publishing of the EvidenceVariable.”

6. for EvidenceVariable.copyrightLabel - correct the typo of resctricted to restricted

7. for EvidenceVariable.handling
Add 'boolean' to the short list.
Add to the bound ValueSet (and CodeSystem)
code = boolean
display = boolean variable
definition = A boolean variable is a variable which is defined to have only 2 possible values: True or False.

8. for EvidenceVariable.category - change short/definition to "A grouping for dichotomous, ordinal, or polychotomouos variables."

9. for EvidenceVariable.conditional
change definition to “The context, situation, or parameters that determine whether the data is obtained to determine the value of the variable.”
change short to “Condition determining whether the data will be collected”

10. for EvidenceVariable.dataStorage - change short to "How the data element (value of the variable) is found"

11. for EvidenceVariable.dataStorage.path - change short to "Where to find the data element in the dataset"
12. for EvidenceVariable.period - change short to "Calendar-based timing when the variable is observed"
13. for EvidenceVariable.constraint.conditional
change definition to “The context, situation, or parameters that determine whether this constraint applies.”
change short to “Condition determining whether this constraint applies”

14. for EvidenceVariable.constraint.conditional.earliestDateTime - change definition to “The earliest permissible value of the variable, used with variables that have a date-based or dateTime-based datatype.”

15. for EvidenceVariable.constraint.conditional.latestDateTime - change definition to “The latest permissible value of the variable, used with variables that have a date-based or dateTime-based datatype.”

for EvidenceVariable Introduction – Scope and Usage
change “PICO (and its variants like PECO or PICOT)” to “PICO (and its variants like PECO using E for Exposure or PICOT adding T for Timing)”
add at the end:  “The EvidenceVariable Resource is also used within ResearchStudy __ to describe the endpoint for an outcome measure.”
for EvidenceVariable Introduction – Boundaries and Relationships
add “The Resource describing a research study protocol (ResearchStudy) will reference EvidenceVariable to describe the endpoint for an outcome measure.”

18. for 5Ws - add author

19. for EvidenceVariable.missingDataMeaning – change cardinality from 0..1 to 0..*

20. for EvidenceVariable.actual – add missing meaning is 'Value unknown.'

21. for EvidenceVariable.constraint.anyValueAllowed - add missing meaning is 'presumed false'