J#49687 Improve the summary of Evidence and Statistics page
As part of CDS R6 prep review, we suggest the following changes to the 'Evidence and Statistics' page:

in 14.7.2 Resources:
a. change EvidenceVariable paragraph to:  The EvidenceVariable Resource provides a machine-interpretable expression of a description of a single evidence variable (e.g., a single exposure or a single outcome), including its precise definition, how the data values are measured, processed, and stored. The definition of an EvidenceVariable may be expressed using a CodeableConcept datatype and/or using a Reference datatype to a Group Resource for a definition combining two or more CodeableConcepts. The EvidenceVariable Resource enables machine-to-machine communication of descriptions and attributes of evidence variables and can be used to find Evidence Resources that match the context of interest, and can be used to understand how to interpret datasets.
b. change Group paragraph to: The Group Resource provides a machine-interpretable expression of a group, and may be expressed as an enumerated group, a definitional group, or a conceptual group. An enumerated group may be specified as a list of members and/or a set of characteristics, and may have a quantity. A definitional group is expressed as a set of characteristics which defines the eligibility criteria for group membership. In FHIR, an enumerated group and a definitional group are each expected to specify the type of group member with allowed types limited to concepts expressed in FHIR Resources. A conceptual group is a group of characteristics that collectively represent a compound definition and may or may not be restricted to use with allowed types. The Group Resource is used within the context of an Evidence Resource to report the evidence variable in the role of population and is also used for providing complex definitions for EvidenceVariable Resources. The Group Resource enables machine-to-machine communication of descriptions and characteristics of groups and can be used to find Evidence Resources that match the population of interest.
c. remove EvidenceReport Resource paragraph
d. change ‘An Evidence-Based Medicine Implementation Guide is in development.’ to ‘An Evidence-Based Medicine on FHIR Implementation Guide is in development.’
in section 14.7.3
CHANGE “In the statistically-defined evidence variables, represented by codeable concepts in Evidence.variableDefinition.variableRole, Population is coded as 'population' or 'subpopulation', Intervention is coded as 'exposure', Comparator is coded as 'referenceExposure', and Outcome is coded as 'measuredVariable'.”
TO “In the statistically-defined evidence variables, represented by codes in Evidence.variableDefinition.variableRole, Population is coded as 'population', Intervention and Comparator are coded as 'exposure', and Outcome is coded as 'outcome'.”
in section 14.7.4 - delete EvidenceReport

add section 14.7.5 Compositions of evidence-related resources:
Reporting a collection of Citation, Evidence, EvidenceVariable, Group, and ArtifactAssessment Resources can be done with a Bundle or List Resource. However, neither Bundle nor List conveys any organization of relationships among the referenced Resources.  Reporting collections of Resources for scientific knowledge typically requires data on their relationships and this is best served with a Composition Resource.

The primary content of a Composition Resource is mainly a repeatable hierarchical section element. Each section element instance may contain a title element (string), code element  (CodeableConcept), and/or focus element  (Reference) to specify what the section is about; a text element (Narrative) for readable summary of the section content; an array of entry elements (Reference) to reference other Resources as the section content; and an array of section elements to contain other sections as the section content.

To support many common patterns of Composition use for EBM communication, the Evidence Based Medicine on FHIR Implementation Guide contains numerous Profiles of Composition.

The EvidenceReport Resource in prior versions of FHIR has been replaced with an EvidenceReport Profile of Composition Resource. See the Evidence Based Medicine on FHIR Implementation Guide for details.