First pass at example
Example for the Observation resource:
Metrics captured from a medical device. The patient-attached device is the subject, the specific device metric is the focus, and the observing device itself is referenced in .device. Specifically, we will use an IV pump as the device and subject, and the focus is the flow rate (device metric). Also include a DeviceAssociation to a Patient.